Changed the process for moving media items into a folder from copy/paste to listing the folders; changed CustomTreeWidget.delete_item to handle multiple selected items; added an "Empty Folder" option to QTreeWidgetItem's context menu.

+++ src/core/projectOn.py
@@ -1,7 +1,7 @@
 """
 This file and all files contained within this distribution are parts of the ProjectOn worship projection software.
 
-ProjectOn v.1.10.0.004
+ProjectOn v.1.10.0.005
 Written by Jeremy G Wilson
 
 ProjectOn is free software: you can redistribute it and/or
@@ -201,7 +201,7 @@ class ProjectOn(QObject):
                 160, 160, Qt.AspectRatioMode.IgnoreAspectRatio, Qt.TransformationMode.SmoothTransformation))
         icon_layout.addWidget(icon_label)
 
-        version_label = QLabel('v.1.10.0.004')
+        version_label = QLabel('v.1.10.0.005')
         version_label.setStyleSheet('color: white')
         version_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
         icon_layout.addWidget(version_label, Qt.AlignmentFlag.AlignCenter)
@@ -904,46 +904,50 @@ class ProjectOn(QObject):
                 connection.close()
             return -1
 
-    def delete_item(self, item: QTreeWidgetItem):
+    def delete_items_from_db(self, items: list[QTreeWidgetItem]):
         """
         Provides a method of deleting a given item from the program's database.
         :param QListWidgetItem item: The item to be removed
         """
         connection = None
         try:
-            if item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'song':
-                table = 'songs'
-                title = item.data(0, Qt.ItemDataRole.UserRole)['title']
-                column = 'title'
-            elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'custom':
-                table = 'customSlides'
-                title = item.data(0, Qt.ItemDataRole.UserRole)['title']
-                column = 'title'
-            elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'image':
-                table = 'imageThumbnails'
-                title = item.data(0, Qt.ItemDataRole.UserRole)['title']
-                column = 'filename'
-                os.remove(self.image_dir + '/' + title)
-            elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'video':
-                file_name = item.data(0, Qt.ItemDataRole.UserRole)['file_name']
-                os.remove(self.video_dir + '/' + file_name)
-                filename_split = file_name.split('.')
-                thumbnail_filename = '.'.join(filename_split[:len(filename_split) - 1]) + '.jpg'
-                os.remove(self.video_dir + '/' + thumbnail_filename)
-
-                table = 'videos'
-                title = item.data(0, Qt.ItemDataRole.UserRole)['title']
-                column = 'filename'
-            elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'web':
-                table = 'web'
-                title = item.data(0, Qt.ItemDataRole.UserRole)['title']
-                column = 'title'
-            else:
-                return
-
             connection = sqlite3.connect(self.database)
             cursor = connection.cursor()
-            cursor.execute(f'DELETE FROM {table} WHERE {column}="{title}"')
+            for item in items:
+                no_command = False
+                table, column, title = '', '', ''
+                if item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'song':
+                    table = 'songs'
+                    title = item.data(0, Qt.ItemDataRole.UserRole)['title']
+                    column = 'title'
+                elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'custom':
+                    table = 'customSlides'
+                    title = item.data(0, Qt.ItemDataRole.UserRole)['title']
+                    column = 'title'
+                elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'image':
+                    table = 'imageThumbnails'
+                    title = item.data(0, Qt.ItemDataRole.UserRole)['title']
+                    column = 'filename'
+                    os.remove(self.image_dir + '/' + title)
+                elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'video':
+                    file_name = item.data(0, Qt.ItemDataRole.UserRole)['file_name']
+                    os.remove(self.video_dir + '/' + file_name)
+                    filename_split = file_name.split('.')
+                    thumbnail_filename = '.'.join(filename_split[:len(filename_split) - 1]) + '.jpg'
+                    os.remove(self.video_dir + '/' + thumbnail_filename)
+
+                    table = 'videos'
+                    title = item.data(0, Qt.ItemDataRole.UserRole)['title']
+                    column = 'filename'
+                elif item.data(0, Qt.ItemDataRole.UserRole)['type'] == 'web':
+                    table = 'web'
+                    title = item.data(0, Qt.ItemDataRole.UserRole)['title']
+                    column = 'title'
+                else:
+                    no_command = True
+
+                if not no_command:
+                    cursor.execute(f'DELETE FROM {table} WHERE {column}="{title}"')
             connection.commit()
             connection.close()
